Make the plugin respect the g:dash_map variable.
[Problem]
For some reason the documentation states that as a user I can set g:dash_map
global variable with a dictionary of "vim filetype" -> "dash docset" and
dash.vim will honor this map when calling Dash.app. Unfortunately this global
variable is never used in the plugin.

[Solution]
On autoload/dash.vim read the global variable and update the plugin map to honor
it.

[Note]
This change also makes possible to define a full list as a value on the
dictionary allowing the user to fully replace all the docsets defined by
default.
Also, the documentation was updated to reflect the change.